Fix and chores
bugfix: StoraygeGroupAllList breaks when json is null
chore: updated tests to follow new naming convention based on Cabinet tests
ui: pull to refresh barebones on ListScreen
