+++ lib/features/cabinet/data/datasources/cabinet_remote_datasource.dart
@@ -30,28 +30,36 @@ class CabinetRemoteDataSource implements ICabinetRemoteDataSource {
               .then((snapshot) => snapshot.data());
 
       if (retrievedSGAllListJson != null) {
-        final StoraygeGroup storaygeGroup =
-            StoraygeGroupAllList.fromJson(retrievedSGAllListJson);
+        if (retrievedSGAllListJson != {"sgSnippetList": []}) {
+          final StoraygeGroup storaygeGroup =
+              StoraygeGroupAllList.fromJson(retrievedSGAllListJson);
 
-        final List<StoraygeGroupSnippet> sgSnippetAllList =
-            storaygeGroup.maybeMap(
-          (_) {
-            throw Exception(
-              "Expected a StoraygeGroupAllList but received the default StoraygeGroup",
-            );
-          },
-          storaygeGroupAllList: (storaygeGroupAllList) =>
-              storaygeGroupAllList.sgAllList,
-          orElse: () {
-            throw Exception(
-              "The variable sgSnippetAllList is null, this could mean that the [sgAllList] document within the [management] collection does not exist.",
-            );
-          },
-        );
+          final List<StoraygeGroupSnippet> sgSnippetAllList =
+              storaygeGroup.maybeMap(
+            (_) {
+              throw Exception(
+                "Expected a StoraygeGroupAllList but received the default StoraygeGroup",
+              );
+            },
+            storaygeGroupAllList: (storaygeGroupAllList) =>
+                storaygeGroupAllList.sgAllList,
+            orElse: () {
+              throw Exception(
+                "This should not happen check back at CabinetRemoteDatasource => getAllListSGSnipFromRemote.",
+              );
+            },
+          );
 
-        return sgSnippetAllList;
+          return sgSnippetAllList;
+        } else {
+          final List<StoraygeGroupSnippet> sgSnippetListEmpty = [];
+
+          return sgSnippetListEmpty;
+        }
       } else {
-        throw FirebaseException(plugin: FS_PLUGIN, message: "Document is null");
+        final List<StoraygeGroupSnippet> sgSnippetListEmpty = [];
+
+        return sgSnippetListEmpty;
       }
     } on FirebaseException catch (e) {
       throw FirebaseException(
