+++ test/core/fixtures/storayge_group_snippet_list_zero.json
@@ -1,3 +1,3 @@
 {
-    "sgSnippet" : [{}]
+    "sgSnippet" : []
 }
