+++ test/features/cabinet/data/datasources/cabinet_remote_datasource_test.dart
@@ -7,6 +7,7 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:storayge/core/constants/app_const.dart';
 import 'package:storayge/features/cabinet/data/datasources/cabinet_remote_datasource.dart';
 import 'package:mocktail/mocktail.dart';
+import 'package:storayge/features/cabinet/domain/entities/storayge_group.dart';
 
 import '../../../../core/fixtures/fixture_reader.dart';
 import '../../../../presets/entities_presets.dart';
@@ -14,6 +15,9 @@ import '../../../../presets/failures_exceptions_presets.dart';
 
 Map<String, dynamic> tAllListSGSnipOneJSON =
     json.decode(fixture('storayge_group_snippet_list_one.json'));
+
+Map<String, dynamic> tAllListSGSnipZeroJSON =
+    json.decode(fixture('storayge_group_snippet_list_zero.json'));
 
 class MockFirebaseFirestore extends Mock implements FirebaseFirestore {}
 
@@ -45,7 +49,7 @@ void main() {
     );
   });
 
-  void setupSuccesfullQuery(Map<String, dynamic> jsonToReturn) {
+  void setupSuccesfullQuery(Map<String, dynamic>? jsonToReturn) {
     when(() => mockFirebaseFirestore.collection(any()))
         .thenAnswer((_) => mockCollectionReference);
     when(() => mockCollectionReference.doc(any()))
@@ -97,6 +101,32 @@ void main() {
               .getAllListSGSnipFromRemote(uid: tUid);
           // assert
           expect(result, tStoraygeGroupAllListSnipOne);
+        },
+      );
+
+      test(
+        'should return an empty List<StoraygeGroupSnippet> when the document retrieved is null',
+        () async {
+          // arrange
+          setupSuccesfullQuery(null);
+          // act
+          final result = await cabinetRemoteDataSource
+              .getAllListSGSnipFromRemote(uid: tUid);
+          // assert
+          expect(result, <StoraygeGroupSnippet>[]);
+        },
+      );
+
+      test(
+        'should return an empty List<StoraygeGroupSnippet> when the sgSnippet array is empty',
+        () async {
+          // arrange
+          setupSuccesfullQuery(tAllListSGSnipZeroJSON);
+          // act
+          final result = await cabinetRemoteDataSource
+              .getAllListSGSnipFromRemote(uid: tUid);
+          // assert
+          expect(result, <StoraygeGroupSnippet>[]);
         },
       );
